refactor: migrate sound_generator
