+++ front/lib/actions/mcp_internal_actions/servers/elevenlabs/utils.ts
@@ -1,38 +1,13 @@
 import { ElevenLabsClient } from "@elevenlabs/elevenlabs-js";
 import { ElevenLabsEnvironment } from "@elevenlabs/elevenlabs-js/environments";
 
+import type {
+  VoiceGender,
+  VoiceLanguage,
+  VoiceUseCase,
+} from "@app/lib/api/actions/servers/speech_generator/metadata";
 import { config as regionsConfig } from "@app/lib/api/regions/config";
 import { dustManagedCredentials } from "@app/types";
-
-export const VOICE_GENDERS = ["female", "male"] as const;
-
-export type VoiceGender = (typeof VOICE_GENDERS)[number];
-
-export const VOICE_LANGUAGES = [
-  "english_american",
-  "english_british",
-  "french",
-  "german",
-  "dutch",
-  "italian",
-  "japanese",
-  "hindi",
-  "chinese",
-] as const;
-
-export type VoiceLanguage = (typeof VOICE_LANGUAGES)[number];
-
-export const VOICE_USE_CASES = [
-  "narrative_story",
-  "conversational",
-  "characters_animation",
-  "social_media",
-  "entertainment_tv",
-  "advertisement",
-  "informative_educational",
-] as const;
-
-export type VoiceUseCase = (typeof VOICE_USE_CASES)[number];
 
 interface VoiceDefinition {
   voiceId: string;
